Implement screen splitting
A user can tick a box that will split the width of the screen into two and draw
copies of everything with an offset. This way the experimenter can also see
what is being drawn on one of their monitor.

Requirement:
  - the Screen('OpenWindow', ...) command has to report width for both
    screens (summed)
  - In order for the size of the grating to be correct, the screen width (in
    cm) must be correctly set up for JUST ONE screen not both.

TODO: test on actual external screens in the lab